Primitives Processing

+++ packages/build-tools/orchestrator/src/process.ts
@@ -12,6 +12,7 @@ import { abortControllers } from "./start.ts";
 export type ProcessComponent = {
   abortController: AbortController;
   process: Deno.ChildProcess;
+  is_piped: { stderr: boolean; stdout: boolean };
   component: ValueOf<typeof ComponentNames>;
   args: string[];
   alive: boolean;
@@ -32,14 +33,33 @@ const wait = (n: number) =>
   new Promise<void>((resolve) => setTimeout(resolve, n));
 
 let shutdownCalled = false;
-export async function shutdown(exitCode: number = 0): Promise<void> {
+export async function shutdown(
+  exitCode: number = 0,
+  errorMessage?: string,
+): Promise<void> {
   if (shutdownCalled) {
     return;
   }
+  if (errorMessage) {
+    console.error(errorMessage);
+  }
+
   shutdownCalled = true;
-  // We are shutting down the processes, so redirect logs to the stdout
-  // So we see the shutdown messages.
-  setCurrentOutput(["otel", "stdout"]);
+  // This allows for any error logs of different processes to be printed, before we terminate them.
+  const timer = {
+    passedTime: 0,
+    waitTime: 2000,
+  };
+  const message = setInterval(() => {
+    console.log(
+      `shutdown called... (waiting ${timer.waitTime - timer.passedTime}ms)`,
+    );
+    timer.passedTime += 300;
+  }, 300);
+  await wait(timer.waitTime);
+  clearInterval(message);
+  console.log("shutdown now");
+
   abortControllers.system.abort();
   abortControllers.noncritical.abort();
   await awaitShutdown();
@@ -54,14 +74,23 @@ async function awaitShutdown(): Promise<void> {
   }
 
   for (const p of processes) {
-    if (p.alive) {
+    if (p.alive && p.component !== ComponentNames.TMUX) {
       console.log("Force killing process", p.process.pid);
       p.process.kill();
     }
   }
 
   await Promise.all(
-    processes.map((process) => process.process[Symbol.asyncDispose]()),
+    processes.filter((p) => p.component !== ComponentNames.TMUX)
+      .map((process) => {
+        if (process.is_piped.stderr) {
+          process.process.stderr.cancel();
+        }
+        if (process.is_piped.stdout) {
+          process.process.stdout.cancel();
+        }
+        process.process[Symbol.asyncDispose]();
+      }),
   );
 }
 
@@ -99,6 +128,10 @@ export const $ = (params: {
     alive: true,
     date: new Date().toISOString(),
     component: params.component,
+    is_piped: {
+      stderr: params.stderr === "piped",
+      stdout: params.stdout === "piped",
+    },
   };
   processes.push(processComponent);
 
@@ -144,10 +177,14 @@ export const $ = (params: {
           return;
         }
 
-        if (!shutdownCalled) {
-          console.error("Shutdown caused by ", params.args);
-        }
-        shutdown(1);
+        shutdown(
+          1,
+          shutdownCalled
+            ? ""
+            : `Shutdown caused by ${params.args.join(" ")}, status ${
+              JSON.stringify(status)
+            }`,
+        );
       }
       failed = true;
     }
